SysTimer, SIM900 UART RX and delay module debugged

+++ src/delay.h
@@ -1,5 +1,5 @@
 #include "io430.h"
 #include "types.h"
 
-u32 Delay_Init(void);
+u8 Delay_Init(void);
 void Delay_DelayMs(u32 value);
